+++ core/parse-xml.c
@@ -697,31 +697,6 @@ static void try_to_match_autogroup(const char *name, char *buf)
 	nonmatch("autogroup", name, buf);
 }
 
-void add_gas_switch_event(struct dive *dive, struct divecomputer *dc, int seconds, int idx)
-{
-	/* sanity check so we don't crash */
-	if (idx < 0 || idx >= dive->cylinders.nr) {
-		report_error("Unknown cylinder index: %d", idx);
-		return;
-	}
-	/* The gas switch event format is insane for historical reasons */
-	struct gasmix mix = get_cylinder(dive, idx)->gasmix;
-	int o2 = get_o2(mix);
-	int he = get_he(mix);
-	struct event *ev;
-	int value;
-
-	o2 = (o2 + 5) / 10;
-	he = (he + 5) / 10;
-	value = o2 + (he << 16);
-
-	ev = add_event(dc, seconds, he ? SAMPLE_EVENT_GASCHANGE2 : SAMPLE_EVENT_GASCHANGE, 0, value, "gaschange");
-	if (ev) {
-		ev->gas.index = idx;
-		ev->gas.mix = mix;
-	}
-}
-
 static void get_cylinderindex(char *buffer, uint8_t *i, struct parser_state *state)
 {
 	*i = atoi(buffer);
